Streamline typings and do not duplicate typings
